full crud on validations controller succesfully

+++ back-end/db/schema.sql
@@ -8,7 +8,7 @@ DROP TABLE IF EXISTS transactions;
 CREATE TABLE transactions (
   id serial PRIMARY KEY,
   item_name varchar(255) NOT NULL,
-  amount integer NOT NULL,
+  amount money NOT NULL,
   date date NOT NULL,
   sender varchar(255) NOT NULL,
   category varchar(255) NOT NULL
